Behavior Testing Update
Adding actions seems to work fine, however, when creating a behavior
with conditions there seems to be a problem constructing the behavior
object. In the constructor, the part with the conditions seems to have
some sort of bug I think. Im not sure where the problem is, it could be
in the conditions java file somewhere. Looking around to try to figure
it out now.